SW-5703 Add recorded sessions to modules
Add a fourth module event type, "Recorded Session," which is for sessions where
the content is a prerecorded video rather than a live video call.

The description of the recorded session is imported from an additional column in
the modules spreadsheet.

Since the session is a recording, it doesn't take place during a specific time
period. But we still want to send out notifications when the recording is ready
to view. For consistency with the way the other event types work, the notification
is sent based on the event's start time, but unlike other event types, the end
time is always set to the same value as the start time such that the "event"
is finished as soon as its time arrives.

The notification text is also customized since the wording for other event types
isn't appropriate for recorded sessions.