SW-5795 Use dynamic organization IDs in tests
As a step toward being able to run tests in parallel, eliminate all the
hardwired organization IDs from tests, with the exception of `PermissionTest`
which will be updated separately.

This is a somewhat large change because a lot of tests used to use the default
organization ID from `DatabaseBackedTest`, which no longer exists. All the
insert methods that take organization ID arguments now default to using the
most recently inserted organization ID instead; in some places this means we
can get rid of explicit organization ID arguments at call sites.

The organization ID arguments of the various insert methods are now required
to be of type `OrganizationId` since there's no longer a valid reason to use
numeric literals.