SW-5697 Add required flag to variables
Add a new column to the all-variables manifest to control whether or not a
given variable is required to be set when it appears in a questionnaire
deliverable.

Currently, nothing enforces this requirement on the server side. The flag
is included in the variable API payload so the web app can use it in the UI.