SW-5761 Restrict plot selection to requested subzones (#2246)
If an observation has a list of requested subzones, only pick monitoring plots in
those subzones when starting the observation.

This also updates how temporary plots are allocated to different subzones.
Previously, priority was given to planted subzones, but now it's given to subzones
that are both planted and requested (if the observation has requested subzones at
all).