SW-5795 Use dynamic IDs in PermissionTest
As a step toward being able to run tests in parallel, update PermissionTest to
stop inserting entities with fixed IDs into the database.

The IDs in this test are meaningful (e.g., all the IDs whose first digit is "1"
belong to the first organization), so the approach here is to map the meaningful
IDs from the test code to underlying IDs from the database. Assertion failure
messages continue to refer to the hardwired IDs from the test so that it's clear
which entities are involved.