SW-5795 Unique cohort and participant names in tests
Cohort and participant IDs and names are required to be globally unique, which
causes problems if two tests run concurrently and both try to insert using the
existing "start with a suffix of 1" naming strategy.

For IDs, switch to dynamically-allocated ones and remove the ability to insert
hardwired IDs.

For names, use random UUID suffixes instead of simple numbers; these should be
sufficiently unique such that the chances of a collision in a single test run
are effectively zero.