SW-5605 Make in-memory test doubles thread-safe
In tests such as controller tests that use Spring-managed dependencies, there is a
single instance of each class that's shared between test threads. This works fine
for stateless classes (which our application code tends to be) but less well for
a couple of in-memory test doubles that allow tests to modify their state.

Update the classes to keep their state in thread-local variables so concurrent
tests won't interfere with each other.