SW-5795 Use dynamic deliverable, module IDs in tests
Stop hardwiring deliverable IDs and module IDs in most tests. Remove the ability
to insert hardwired module IDs.

There is a hardwired deliverable ID in the actual application code: the
deliverable that's associated with the variables that are used for pre-screening
checks. Update the code so that tests can override that ID as needed.

There are also a few remaining hardwired deliverable IDs in the test that tries
to import a copy of the real all-variables file; the deliverables mentioned in
the file need to exist for the import to succeed. This change reduces the risk
of ID collisions by getting rid of one of the two tests that imported that file,
but the possibility of those IDs colliding with dynamically-allocated ones still
exists.

There were a few hardwired project IDs in `DeliverableStoreTest`; those are
removed here as well.