SW-5553 File support ticket when document upload fails
If the Google or Dropbox folders for a project are misconfigured, either in our
database or on the Google/Dropbox side, uploading a document for a document
deliverable will fail, but the accelerator admins will never find out about it
since they don't monitor server errors.

With this change, failed deliverable document uploads will cause support tickets
to be created in Jira; the tickets can then be routed to whichever person is
responsible for configuring the project's folders.