SW-5841 Run variable workflow update as system user (#2287)
Since a regular org user can update a variable value, which consequently updates the variable's workflow details (by setting it to "In Review"), run the update as a system user with elevated permissions.

In order to ensure we know who updated a variable's status, the user ID can optionally be passed in to the update function, otherwise it defaults to the current user.