SW-5754 API to manage organization internal tags
Currently, internal tags can be assigned to organizations via the admin UI, but
not via the REST API. Add endpoints to list the available internal tags, to
list the internal tags for an organization, and to update an organization's
internal tags.

Creation and deletion of custom admin-defined tags continues to be an admin UI
feature, though the new API endpoints can be used to add custom tags to
organizations once they've been created.