Exclude tables from SchemaDocsGenerator (#2305)
Now that we're running tests in parallel, sometimes SchemaDocsGenerator runs after
a test that calls the JobRunr API and causes JobRunr to create its database
tables. This was causing intermittent test failures since those tables aren't
listed in SchemaDocsGenerator.

Add logic to exclude tables from the table list; use it for a few of the existing
tables that are used by libraries instead of our application code.