SW-5754 List all organizations and internal tags
To support editing internal tags for organizations in the web app, the frontend
needs to be able to show a menu of available organizations and which tags they
have. Add an endpoint to return the full list of organizations and each one's
tags, only accessible by super-admin users.

For consistency, this also changes the URL paths of the endpoints introduced in
commit c97c8296 so they all fall under `/api/v1/internalTags`.