SW-5900 Stop requiring map area to match land use totals
The original requirement for pre-screening was to check that the area of the map
boundary was approximately equal to the total of the hectare counts for different
land usage model types, but that turned out to be too finicky. Instead, make sure
both the boundary area and the total of the land usage model type hectares fall
within the minimum and maximum for the site's country.