Remove Spring Boot R2DBC workaround (#2342)
At one point there was a problem with Spring Boot's jOOQ autoconfiguration where
it would try to set up a database library we don't use; as a workaround we added
configuration to explicitly disable that library. The problem has since been
fixed in Spring Boot, so the workaround isn't needed any more.