SW-5926 Create Google Drive folders on demand
Currently we require an admin to configure the Google Drive folder to use for
uploaded documents for projects, but that doesn't work well for applications since
they won't have been reviewed yet at the time they need to start uploading
documents.

Add a config option that specifies a Google Drive folder in which new project
folders will be created on demand for projects that don't currently have folders.