Fix missing foreign keys in migration
The V302 migration succeeded on an empty database (such as the one that's
used to run tests) but failed if the `deliverable_categories` table was
already populated from a previous server run because it was trying to
refer to interest IDs that hadn't been inserted yet.

This is an edit to an existing migration, but shouldn't require any Flyway
migration history surgery because it's either running on a clean database
(in which case the previous version won't have been run anyway) or it will
be running on a database where the previous version failed.