SW-5959 Support backfilling old variable upgrades
We imported new all-variables CSVs in local dev environments and the staging
environment before adding support for automatically upgrading variables,
meaning that there are values of out-of-date variables that haven't been
brought forward to the replacement variables.

Add a button to the admin UI to walk back through the replacement history of
all the variables and upgrade any values that need it.

This should only need to be done once, since future changes to the all-variables
list will trigger automatic upgrades.