SW-5951 Upgrade calculator for all-variables list (#2372)
When a new version of the all-variables list is imported, we want to upgrade
the values of any variables that are being replaced by new versions. The process
is basically the same as the existing logic for upgrading documents to new
variable manifests, but unlike that upgrade process, it applies across all
variables in all projects.

Add the code that calculates the required write operations. Nothing calls this
yet; that'll be added in upcoming changes.