SW-5905 Create HubSpot deal on application submit
When an application is submitted for review for the first time, create information
about it in HubSpot and update the project's details with a link to its HubSpot
deal.

The bulk of this change is to add integration with HubSpot's API. There are three
new server configuration options, and there is also a new authorization flow,
available in the admin UI, that authorizes terraware-server to connect to a
HubSpot account. The credentials from that authorization flow are stored in the
database since they aren't known in advance.

Instructions for configuring a dev environment to connect to a HubSpot sandbox
account are in `docs/HUBSPOT.md`.