Fix tests that assume DAO.findAll result order
The findAll method on the jOOQ DAO classes doesn't return results in any
specific order. Usually, it happens to return results in insertion order, but,
especially when multiple tests are hitting the database concurrently, the order
can be different.

Fix all the tests that compare findAll's return value to a list of multiple
rows without sorting the list; they now use assertTableEquals instead.