Add convenience methods for asserting table contents (#2418)
A common pattern in our tests is to call some application code and then verify
that the database was updated correctly. This usually looks something like

    assertEquals(
        listOf(
            SomeTableRow(col1 = "foo", col2 = "bar"),
        ),
        someTableDao.findAll().map { it.id = null })

This pattern is an occasional source of spurious test failures when we forget
that `findAll()` doesn't guarantee any particular order of results.

Introduce a new set of methods in `DatabaseBackedTest` that are more succinct to
call, offer more flexibility, and avoid the possibility of test failures due to
assumptions about result order.