Limit size of parallel test thread pool
Currently, new threads are added to the thread pool for parallel test execution
over time, which causes the test suite's memory and CPU usage to grow as it runs.
In some cases, such as on CI runners, this can cause the tests to run out of
memory and abort.

Add configuration options to limit the size of the thread pool.