Fix caching of internal-only variables (#2434)
We were using a single cache for all variables, and the check for permission to
read internal-only variables was only happening on cache miss. As soon as a
privileged user fetched deliverable variables, the internal-only variables would
be in the cache and would start being returned to all users regardless of
privilege.

Since there can be internal-only variables nested inside public variables, the fix
isn't as simple as just filtering internal-only variables from the "list
variables" API responses for unprivileged users; the same variable could be
visible to both privileged and unprivileged users but with a different set of
child variables.

Fix the problem by using a separate cache for unprivileged users. It will only
contain public variables, both at the top level and nested in parent variables.