SW-6052 Support getting specific variables
Currently, the APIs for fetching variables and variable values operate in bulk:
they return all the variables for a given deliverable or document.

To support cross-deliverable variable dependencies, where the client needs to
reference the value of a variable that isn't part of the deliverable it's focused
on, add optional `variableId` query parameters to both the variable and variable
value GET endpoints to access variables regardless of which deliverable or
document they're in. Multiple variables may be fetched in one request by
specifying the `variableId` parameter more than once in the URL.