Support regular events in RateLimitedEventPublisher (#2445)
Allow regular events to be published using `RateLimitedEventPublisher` so that
callers that need a mix of immediate and rate-limited publication don't have to
depend on both `ApplicationEventPublisher` and `RateLimitedEventPublisher`.

Rename `publishOrDefer` to `publishEvent` to match the regular event publishing
API; callers shouldn't have to care if the event will be deferred or not.