Upgrade section default value variable references
When a new version of the all-variables list is uploaded, there may be new
versions of variables that are referenced in default section values. Replace
those references with the updated versions of the variables so that newly-created
documents don't start off with references to outdated variables.

This change does not affect references in any existing documents, just the default
ones that are used at document creation time.