SW-5585 Notify completed parent section owners (#2452)
The notification logic in commit 9f22cf7 wasn't quite right: it only notified
owners of parent sections about edits to variables referenced by child sections if
the child sections were marked as completed. But the typical scenario is that the
child section has no completion status at all (because it's unnumbered) and the
parent section is marked as completed.

Change the code to always look at the completion status of parent sections
regardless of the status of the child section that actually has the reference to
the updated variable.