SW-6099 Added feature to specify grid origin when importing planting sites (#2482)
Our plot size detection uses a grid system that is aligned to the bounding box of the planting site boundary. 

For a small bounding site, this can cause issues because the grid alignment may lead to a lack of solutions.

This temporary fix adds the ability for the user to specify a grid origin point (easiest way is to use a SW corner of a valid permanent plot). Screenshot belows show the lat-long to make this work for this site.

In longer term, a more complete solution is to think about ways to automatically compute a grid system that will not split up potential solution (likely some topography/convex hull problem so maybe this is not worth the effort at all).

But given that this is a very small edge case for very small planting sites. This should be okay for now. 

![Screenshot 2024-10-07 at 4.04.09 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/NOxkRPuhAfT4F7nmdXtG/1cb6269b-65b7-4b29-be6b-53e98a2eb3fa.png)