SW-6146 Remove unused list of supported locales
We used to configure Spring with an explicit list of supported locales, but we
removed that in commit 51b589f7 to support adding country codes to locales to get
country-specific date and number formatting.

Remove the list of locales we used to pass to Spring; people might think they
need to edit it when adding languages, when in fact it isn't referenced anywhere.