Fix usages of obsolete javax.inject package
The old `javax.inject` package is `jakarta.inject` nowadays, but we've had some
dependencies that pull in the old version, so it's been easy to pick the wrong one
in the IDE autocomplete menu by mistake.

Switch to the new package name in the places where we were using the outdated one.