SW-6131 Ignore additional deliverables CSV column (#2501)
In preparation for moving the mapping between variables and deliverables to the
deliverables spreadsheet, make the deliverables importer ignore the extra column
where the variables list will appear. This will let us start working on
populating that column in the spreadsheet before the code is updated to start
using it.