SW-6131 Move deliverable positions to new table (#2508)
Add a new `deliverable_variables` table to hold the list of which variables are
associated with which deliverables, including their positions within the
deliverables. This replaces the `deliverable_id` and `deliverable_position`
columns on the `variables` table.

This is a prerequisite for associating variables with multiple deliverables,
but that functionality isn't introduced here. This change should have no
user-visible effect on the system's behavior.