SW-6131 Associate variables from deliverables sheet (#2518)
Make DeliverablesImporter populate the list of variables for each deliverable
if present in the deliverables sheet.

A variable may be associated with more than one deliverable.

If the "Variables" column is empty for a deliverable, the deliverable's existing
list of variables is left untouched; this is to support gradually transitioning
from the old scheme where deliverable IDs were specified on the all-variables
sheet by moving the associations from that sheet one deliverable at a time.