SW-6176 Fix adding species to approved species list
Adding a species to a species list that was already approved was failing if the
user didn't have accelerator admin privileges because it required resetting the
submission status, which was a privileged operation.

Update the permission check to allow resetting submission status to "Not
Submitted" if the user has permission to create a new submission.