SW-6185 Add plot size to database schema
In preparation for supporting different monitoring plot sizes, add a column to the
`monitoring_plots` table to keep track of the size of each plot, and mirror that
value in the monitoring plot model class. For now, the size is always 25 meters,
such that this doesn't change any existing behavior.