SW-6185 Expose monitoring plot size to clients (#2528)
Include the monitoring plot's size in API responses that contain monitoring plot
information, including the search API. This will allow the client to show the
plot size and to calculate statistics that depend on plot size, e.g., the total
area covered by the results of an observation, correctly even when the plot size
changes in the future.

The plot size isn't changed here; it's still always 25 meters.