SW-6187 Add overlapping plots to data model, API (#2531)
Changing the plot sizes for future observation of a planting site that already has
observation results will mean that the new-sized plots will partially overlap the
old-sized ones. We want users to be able to navigate back to the old-sized plots
to see how the observation data has changed over time.

Add the concept of overlapping plots to the system. A newer plot may overlap with
some number of older plots, and may in turn be overlapped by some number of even
newer plots. Include the overlapping plot IDs in the plot-level observation results.

Currently, since plot size isn't variable yet, the list of overlapping plots will
always be empty.