SW-6186 Use plot size in planting density calculation (#2532)
Remove the hardwired plot size from the plot-level planting density calculation in
`ObservationResultsStore` so that the density number is calculated correctly as
the plot size varies.

Add a copy of one of the observation calculation test scenarios with the expected
planting density and estimated total plants adjusted for 30m plots instead of 25m.