SW-6193 Validate that zones can fit 1-plot clusters
Previously, the validation logic for planting zone geometries checked whether
there was room for a 4-plot permanent cluster and a single temporary plot.
Remove the 4-plot cluster check and instead validate that there's room for two
monitoring plots (which will allow an observation with one permanent and one
temporary plot).