SW-6191 Reduce number of temporary plots
The default number of temporary plots in a planting zone is based on the number of
permanent plots, which used to be 4 times the number of permanent clusters but is
now the same as the number of permanent clusters. Update the temporary plot counts
accordingly, and update the planting site editor in the admin UI to base the
temporary plot calculation on the correct value.