SW-6205 Make workflow history follow variable replacements
If a variable is replaced by a newer version, we want the workflow history of the
old variable to be considered part of the history of the new variable too.

This should fix the system appearing to forget the statuses and comments of
variables when an upload of the all-variables CSV causes variables to be replaced.