When variables are upgraded, also replace injections/references to the old variable within section text for documents (#2546)
When a variable is upgraded (because its definition is changed within the all variables CSV and uploaded), there may still be old injections/references to this variable within the section text for documents. This causes a few issues, the most obvious being that the frontend receives multiple versions of a variable. The frontend receives the old version of the variable, because it exists within the value of a variable present (the section variable), and it receives the new variable because that is the current version of a variable that can be injected/referenced within the document.

This PR fixes the issue by adding an additional step to the "Upgrade Old Variable Values" which goes through and upgrades old injections/references to "old" variables.