Admin UI home page improvements
Organize the various parts of the admin UI home page into high-level sections.
The sections start off hidden and can be expanded by clicking on their names.
The expansion status of all the sections is saved in the browser's local storage
so that it's remembered across page loads.

Change the admin UI's title to "Terraware Administration" since it isn't really
a placeholder for anything any more.