SW-6222 Add observed species to subzone species list (#2573)
When the client requests the list of species in a subzone (the mobile app uses
this to populate the list of options when observing a monitoring plot), we
currently return the species that were reported to have been planted in the
subzone.

Start also including all the species that have been observed at the planting site,
even if they were never reported as planted in the subzone. Among other things,
this will cover cases where a plant has spread naturally from one subzone to
another.