SW-6230 Allow fetching of application submissions
When the web app saves changes to fields such as the internal comment on an
application deliverable, it tries to refresh the deliverable data using
`GET /api/v1/accelerator/deliverables/{id}/submissions/{projectId}`. Currently,
that endpoint only supports non-application deliverables (application deliverables
are exposed via an alternate endpoint) because the underlying query code needs
to avoid returning lists of application deliverables on the Deliverables page
in the app.

This doesn't appear to cause any user-visible problems, but it does cause an
error message to be logged by the server.

Since the client is requesting a specific deliverable in this case, there's no
danger of returning a list that mixes application and non-application
deliverables. Update the query to only exclude application deliverables if the
caller didn't supply a specific deliverable ID.