SW-6244 Don't log error if GBIF species not found
If the user searches for a species that's not in the GBIF database, we shouldn't
consider it a server error, just a normal outcome of the request.

Update the GBIF lookup endpoint to throw an explicit web-server "not found"
exception rather than an application-defined exception, and remove the error
logging for that kind of exception since when we throw it, it will always mean
the code is deliberately returning a response to the user.