Detect permission inversions (#2589)
We've had a number of bugs that are caused by permission inversions: situations
where we require a particular permission to perform a function, but then the
function ends up needing to do something that requires a higher level of
permissions, either directly or via an event handler. We sometimes fail to find
these bugs in testing if we happen to test with a user who passes both the
permission checks.

This change adds logic to track which permission checks are guarded by which other
checks. If it detects an inversion while handling a user request, it logs a
warning message. If it detects an inversion in a unit test, it fails the test.
In either case, the output includes abbreviated stack traces of the two checks
to make it easier to fix the problem.

Only tests that implement `RunsAsDatabaseUser` interface added in commit 137c89a2
will have their permission checks tracked and verified.