Check for permission inversions in controller tests
Make the controller integration tests implement `RunsAsDatabaseUser` so they check
for permission inversions during request handling.

This requires a little bit of finagling because a fresh copy of the user is loaded
from the database as part of the request handling code. We need to grab that copy
of the user in order to see what permission checks it recorded.