SW-6260 Add more detail to variable value logs (#2606)
To make it easier to troubleshoot issues with variable updates, log more details
about variable value operations, and do it in a structured way using MDC so it's
searchable using attribute filters in Datadog.

Make tests use the default log format that's used in local dev environments, which
includes the MDC values.