Use separate file for access logs (#2608)
Commit 789b3e00 switched HTTP access logs to use JSON format so we could get
richer context information about requests. It put the access logs in the same
log file as the main server logs. This worked after the server started, but
when logs were rotated, the access logs didn't switch over to the newly-created
server logfile.

Switch to a separate file for access logs. The new file will be rotated on the
same schedule as the main server log.