SW-6251 Add history IDs to observation plots (#2614)
To prepare for the ability to redefine subzones that already contain monitoring
plots, start recording which versions of plots were included in observations.
This will allow us to tell the user which subzone a plot was in at the time of
the observation, even if the site map has subsequently changed.