SW-6162 Fix issue with disappearing variable value references / injections when a document is updated to a new variable manifest version (#2612)
When a document template has a new variable manifest uploaded for it, section variable IDs may be updated, for example if the section name goes from "Project Details" to "The Project Details". When this occurs, a user may "upgrade" their document (which points at the old version of the variable manifest) to the new version. 

During this upgrade, for project values within each section, "update" and "delete" operations were created to determine the upgrade path for the user-entered data. Now that the non-section variables exist outside of the manifest, and any out-dated references to them are updated during their own upgrade, there is no longer a need to create the "delete" operations which were responsible for cleaning up variable references that no longer existed. 

This solves a bug that was deleting references to variables within section text that belonged to sections that were part of an upgraded variable manifest.